[Sema] Synthesize domain of @objc enums properly
The domain constant generated by PrintAsObjC always has the value
`"ModuleName.TypeName"`, but the actual value of the derived
`_NSErrorDomain` constant from the `_BridgedNSError` protocol was
implemented as `"\(self)"`. This meant that any time that the default
string representation of the enum type did not match the format
`"ModuleName.TypeName"`, the generated domain constant did not match the
actual error domain. This could happen when compiling programs at the
command-line that define the error (as the module name is typically
omitted there), or if the enum was nested inside another @objc
declaration (as the qualified type name there includes the parent type
declaration).

Fixes SR-700.