[Constraint solver] Introduce one-way binding constraints.
Introduce the notion of "one-way" binding constraints of the form

  $T0 one-way bind to $T1

which treats the type variables $T0 and $T1 as independent up until
the point where $T1 simplifies down to a concrete type, at which point
$T0 will be bound to that concrete type. $T0 won't be bound in any
other way, so type information ends up being propagated right-to-left,
only.

Experiment with this idea by introducing a new command-line flag,
`-enable-solver-one-way-constraints`, which introduces one-way
constraints for (1) the parameter types of closures, and (2) the
return type of single-expression closures. This means that type
inference will flow into the closure from the parameters, through the
closure body (for a single-expression closure), and then out of the
closure. There is no backward propagation from the use of a closure's
result type through the closure body, so this behavior is mimicking
the unidirectional flow of type inference in a normal function body or
multi-statement closure.

Part of rdar://problem/50150793.