[Constraint graph] Compute "solve plan" based on one-way constraints.
One-way constraints will allow the constraint solver to isolate more
components to solve (more) independently. When there is a one-way
constraint such as "$T5 one-way bind to $T4", the component associated
with $T4 can be solved independently of any of the constraints on
$T5. Once $T4 is solved, $T5 will be bound to its result (e.g., the
same as a "bind" constraint), constraints involving $T5 can be
solved.

Rework the connected components algorithm in the constraint graph to
consider one-way constraints. The algorithm proceeds in several steps:

1. Compute connected components, excluding one-way constraints from
consideration.
2. Compute a directed graph amongst the components using only the
one-way constraints, where an edge A -> B indicates that the type
variables in component A need to be solved before those in component
B.
3. Compute a "solve plan" that describes the order in which the
components need to be solved as well as describing which
(previously-computed) components

At this point, the only meaningful output is the "solve plan", which
is used for testing purposes. The solver does not actually use the
solve plan for anything, so we collapse components tied together by
one-way edges and let everything else continue as before. This is a
staging mechanism that will be eliminated later.