[Type checker] Delay the diagnosis of missing witnesses further.
The handling of "global" missing witnesses would diagnose a missing witness
(showing "almost" matches) and then, later, provide the Fix-It to add
the declaration if the user wants it. Delay the diagnostic part until later,
when we add the missing witness.

While here, correct an existing issue with memory corruption that
occurs because we would diagnose some of these witnesses, then clear a
"global" witness list, while still iterating through parts of that
list.